+++ server/resources/errorFactory.js
@@ -1,22 +1,4 @@
-const Errors = {
-    DATABASE_ERROR: 0,
-    OBJECT_NOT_FOUND: 1,
-    MISSING_FIELD: 2,
-    SHORT_STRING: 3,
-    INVALID_ID: 4,
-    INVALID_DATA_TYPE: 5,
-    INVALID_LENGTH: 6,
-    INVALID_FORMAT: 7,
-    DATE_ORDER: 8,
-    NESTED_OBJECT_NOT_FOUND: 9,
-    LONG_STRING: 10,
-    NUMBER_LOWER_BOUND: 11,
-    NUMBER_UPPER_BOUND: 12,
-    INVALID_NUMBER: 13,
-    INVALID_FIELD: 14,
-    INVALID_ENCODING: 15,
-    NOT_AUTHORIZED: 16
-}
+const Errors = require('./Errors');
 
 const buildError = (error, arg1, arg2) =>{
     var errorObj = {
@@ -106,6 +88,5 @@ const buildError = (error, arg1, arg2) =>{
 }
 
 module.exports = {
-    Errors,
     buildError
 }
