+++ server/resources/index.js
@@ -1,5 +1,6 @@
 const ErrorFactory = require('./errorFactory');
-
+const Errors = require('./Errors').Errors; 
 module.exports = {
-    ErrorFactory
+    ErrorFactory,
+    Errors
 }
